% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoobears_cost.R
\name{zoobears_cost}
\alias{zoobears_cost}
\title{Title ZooBears Cost}
\usage{
zoobears_cost(price, count)
}
\arguments{
\item{price}{price of a zoobear}

\item{count}{number of bears at the zoo}
}
\value{
Returns the cost of total zoobears entered
}
\description{
Title ZooBears Cost
}
